Add payload to state: works for input, but not select
For satisfactory user experience, form values must be
kept intact in case of (validation) errors.

The useStateAction hook, however, resets the form.

We can still achieve the goal with following technique:

- add the form's payload (FormData) to the state
- use the payload values to set the default values
  of the form controls (using the defaultValue
  prop for input of type select and text, and
  defaultChecked for input of type checkbox and radio).

PROBLEM: This works for the text and radio inputs, but
not for select, as described in comment:

[React 19] allow opting out of automatic form reset
https://github.com/facebook/react/issues/29034#issuecomment-2221492687

A 2022 bug report already exists about defaultValue
not working with select after the initial render:

Bug: defaultValue is not consistent between input
and select
https://github.com/facebook/react/issues/24165

The problem can be mitigated with use of a useEffect hook
(see next commit).