payload technique works for input, but not select
payload technique when using useStateAction works for input form
controls (both text and radio), but not for select.

Am experiencing issue described in comment:

[React 19] allow opting out of automatic form reset
https://github.com/facebook/react/issues/29034#issuecomment-2221492687

A bug report already exists about defaultValue not working right
with select:

Bug: defaultValue is not consistent between input and select
https://github.com/facebook/react/issues/24165
sync

In the next commit, I'll use a useEffect hook to workaround this temporarily until that React bug is fixed.